+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/v2/ui/workorder/WorkOrderScreen.java
@@ -1089,12 +1089,12 @@ public class WorkOrderScreen extends RelativeLayout implements UUIDView {
     /*-*****************************-*/
     private final DocumentClient _documentClient = new DocumentClient() {
         @Override
-        public boolean processDownload(long documentId) {
+        public boolean processDownload(int documentId) {
             return true;
         }
 
         @Override
-        public void onDownload(long documentId, File file, int state, boolean isSync) {
+        public void onDownload(int documentId, File file, int state, boolean isSync) {
             Log.v(TAG, "DocumentClient.onDownload");
             if (file == null || state == DocumentConstants.PARAM_STATE_START) {
                 if (state == DocumentConstants.PARAM_STATE_FINISH)
